Remove 'New' naming from Zustand architecture - replace old implementations
