+++ packages/services/src/ui/context/OxyContext.tsx
@@ -1,10 +1,15 @@
-import React, { createContext, useContext, useState, useEffect, useCallback, ReactNode, useMemo } from 'react';
+/**
+ * OxyContext using Zustand stores
+ * This replaces the complex Redux + Context combination with a clean, performant solution
+ */
+
+import React, { createContext, useContext, useEffect, ReactNode, useRef } from 'react';
 import { OxyServices } from '../../core';
 import { User } from '../../models/interfaces';
-import { SecureLoginResponse, SecureClientSession, MinimalUserData } from '../../models/secureSession';
-import { DeviceManager } from '../../utils/deviceManager';
+import { SecureClientSession, MinimalUserData } from '../../models/secureSession';
+import { initializeOxyStore, useAuth, useFollow } from '../../stores';
 
-// Define the context shape
+// Define the context shape - maintaining backward compatibility
 export interface OxyContextState {
   // Authentication state
   user: User | null; // Current active user (loaded from server)
@@ -44,724 +49,153 @@ export interface OxyContextState {
   hideBottomSheet?: () => void;
 }
 
-// Create the context with default values
+// Create the context
 const OxyContext = createContext<OxyContextState | null>(null);
 
-// Props for the OxyContextProvider
+// Props for the OxyContextProvider - maintaining backward compatibility
 export interface OxyContextProviderProps {
   children: ReactNode;
   oxyServices: OxyServices;
-  storageKeyPrefix?: string;
-  onAuthStateChange?: (user: User | null) => void;
+  storageKeyPrefix?: string; // Kept for backward compatibility but not used
+  onAuthStateChange?: (user: User | null) => void; // Kept for backward compatibility
   bottomSheetRef?: React.RefObject<any>;
+  showBottomSheet?: (screenOrConfig?: string | { screen: string; props?: Record<string, any> }) => void;
+  hideBottomSheet?: () => void;
 }
 
-// Platform storage implementation
-interface StorageInterface {
-  getItem: (key: string) => Promise<string | null>;
-  setItem: (key: string, value: string) => Promise<void>;
-  removeItem: (key: string) => Promise<void>;
-  clear: () => Promise<void>;
-}
-
-// Web localStorage implementation
-class WebStorage implements StorageInterface {
-  async getItem(key: string): Promise<string | null> {
-    return localStorage.getItem(key);
-  }
-
-  async setItem(key: string, value: string): Promise<void> {
-    localStorage.setItem(key, value);
-  }
-
-  async removeItem(key: string): Promise<void> {
-    localStorage.removeItem(key);
-  }
-
-  async clear(): Promise<void> {
-    localStorage.clear();
-  }
-}
-
-// React Native AsyncStorage implementation
-let AsyncStorage: StorageInterface;
-
-// Determine the platform and set up storage
-const isReactNative = (): boolean => {
-  return typeof navigator !== 'undefined' && navigator.product === 'ReactNative';
-};
-
-// Get appropriate storage for the platform
-const getStorage = async (): Promise<StorageInterface> => {
-  if (isReactNative()) {
-    if (!AsyncStorage) {
-      try {
-        const asyncStorageModule = await import('@react-native-async-storage/async-storage');
-        AsyncStorage = asyncStorageModule.default;
-      } catch (error) {
-        console.error('Failed to import AsyncStorage:', error);
-        throw new Error('AsyncStorage is required in React Native environment');
-      }
-    }
-    return AsyncStorage;
-  }
-
-  return new WebStorage();
-};
-
-// Storage keys for secure sessions
-const getSecureStorageKeys = (prefix = 'oxy_secure') => ({
-  sessions: `${prefix}_sessions`, // Array of SecureClientSession objects
-  activeSessionId: `${prefix}_active_session_id`, // ID of currently active session
-});
-
+/**
+ * OxyContextProvider using Zustand stores
+ * Initializes the Zustand store and provides OxyServices access
+ */
 export const OxyContextProvider: React.FC<OxyContextProviderProps> = ({
   children,
   oxyServices,
-  storageKeyPrefix = 'oxy_secure',
-  onAuthStateChange,
+  storageKeyPrefix, // Kept for backward compatibility
+  onAuthStateChange, // Kept for backward compatibility
   bottomSheetRef,
+  showBottomSheet,
+  hideBottomSheet,
 }) => {
-  // Authentication state
-  const [user, setUser] = useState<User | null>(null);
-  const [minimalUser, setMinimalUser] = useState<MinimalUserData | null>(null);
-  const [sessions, setSessions] = useState<SecureClientSession[]>([]);
-  const [activeSessionId, setActiveSessionId] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [storage, setStorage] = useState<StorageInterface | null>(null);
-
-  // Storage keys (memoized to prevent infinite loops)
-  const keys = useMemo(() => getSecureStorageKeys(storageKeyPrefix), [storageKeyPrefix]);
-
-  // Initialize storage
+  const isInitialized = useRef(false);
+  const onAuthStateChangeRef = useRef(onAuthStateChange);
+  
+  // Update the ref when the callback changes
   useEffect(() => {
-    const initStorage = async () => {
-      try {
-        const platformStorage = await getStorage();
-        setStorage(platformStorage);
-      } catch (error) {
-        console.error('Failed to initialize storage:', error);
-        setError('Failed to initialize storage');
-      }
-    };
+    onAuthStateChangeRef.current = onAuthStateChange;
+  }, [onAuthStateChange]);
 
-    initStorage();
-  }, []);
-
-  // Effect to initialize authentication state
+  // Initialize the store once
   useEffect(() => {
-    const initAuth = async () => {
-      if (!storage) return;
-
-      setIsLoading(true);
-      try {
-        // Load stored sessions
-        const sessionsData = await storage.getItem(keys.sessions);
-        const storedActiveSessionId = await storage.getItem(keys.activeSessionId);
-
-        console.log('SecureAuth - sessionsData:', sessionsData);
-        console.log('SecureAuth - activeSessionId:', storedActiveSessionId);
-
-        if (sessionsData) {
-          const parsedSessions: SecureClientSession[] = JSON.parse(sessionsData);
-
-          // Migrate old session format to include user info
-          const migratedSessions: SecureClientSession[] = [];
-          let shouldUpdateStorage = false;
-
-          for (const session of parsedSessions) {
-            if (!session.userId || !session.username) {
-              // Session is missing user info, try to fetch it
-              try {
-                const sessionUser = await oxyServices.getUserBySession(session.sessionId);
-                migratedSessions.push({
-                  ...session,
-                  userId: sessionUser.id,
-                  username: sessionUser.username
-                });
-                shouldUpdateStorage = true;
-                console.log(`Migrated session ${session.sessionId} for user ${sessionUser.username}`);
-              } catch (error) {
-                // Session might be invalid, skip it
-                console.log(`Removing invalid session ${session.sessionId}:`, error);
-                shouldUpdateStorage = true;
-              }
-            } else {
-              // Session already has user info
-              migratedSessions.push(session);
-            }
-          }
-
-          // Update storage if we made changes
-          if (shouldUpdateStorage) {
-            await saveSessionsToStorage(migratedSessions);
-          }
-
-          setSessions(migratedSessions);
-
-          if (storedActiveSessionId && migratedSessions.length > 0) {
-            const activeSession = migratedSessions.find(s => s.sessionId === storedActiveSessionId);
-
-            if (activeSession) {
-              console.log('SecureAuth - activeSession found:', activeSession);
-
-              // Validate session
-              try {
-                const validation = await oxyServices.validateSession(activeSession.sessionId);
-
-                if (validation.valid) {
-                  console.log('SecureAuth - session validated successfully');
-                  setActiveSessionId(activeSession.sessionId);
-
-                  // Get access token for API calls
-                  const tokenResponse = await oxyServices.getTokenBySession(activeSession.sessionId);
-                  // Set the token on the service instance
-                  oxyServices.setTokens(tokenResponse.accessToken, '');
-
-                  // Load full user data
-                  const fullUser = await oxyServices.getUserBySession(activeSession.sessionId);
-                  setUser(fullUser);
-                  setMinimalUser({
-                    id: fullUser.id,
-                    username: fullUser.username,
-                    avatar: fullUser.avatar
-                  });
-
-                  if (onAuthStateChange) {
-                    onAuthStateChange(fullUser);
-                  }
-                } else {
-                  console.log('SecureAuth - session invalid, removing');
-                  await removeInvalidSession(activeSession.sessionId);
-                }
-              } catch (error) {
-                console.error('SecureAuth - session validation error:', error);
-                await removeInvalidSession(activeSession.sessionId);
-              }
-            }
-          }
-        }
-      } catch (err) {
-        console.error('Secure auth initialization error:', err);
-        await clearAllStorage();
-      } finally {
-        setIsLoading(false);
-      }
-    };
-
-    if (storage) {
-      initAuth();
+    if (!isInitialized.current) {
+      initializeOxyStore(oxyServices);
+      isInitialized.current = true;
     }
-  }, [storage, oxyServices, keys, onAuthStateChange]);
-
-  // Remove invalid session
-  const removeInvalidSession = useCallback(async (sessionId: string): Promise<void> => {
-    const filteredSessions = sessions.filter(s => s.sessionId !== sessionId);
-    setSessions(filteredSessions);
-    await saveSessionsToStorage(filteredSessions);
-
-    // If there are other sessions, switch to the first one
-    if (filteredSessions.length > 0) {
-      await switchToSession(filteredSessions[0].sessionId);
-    } else {
-      // No valid sessions left
-      setActiveSessionId(null);
-      setUser(null);
-      setMinimalUser(null);
-      await storage?.removeItem(keys.activeSessionId);
-
-      if (onAuthStateChange) {
-        onAuthStateChange(null);
-      }
-    }
-  }, [sessions, storage, keys, onAuthStateChange]);
-
-  // Save sessions to storage
-  const saveSessionsToStorage = useCallback(async (sessionsList: SecureClientSession[]): Promise<void> => {
-    if (!storage) return;
-    await storage.setItem(keys.sessions, JSON.stringify(sessionsList));
-  }, [storage, keys.sessions]);
-
-  // Save active session ID to storage
-  const saveActiveSessionId = useCallback(async (sessionId: string): Promise<void> => {
-    if (!storage) return;
-    await storage.setItem(keys.activeSessionId, sessionId);
-  }, [storage, keys.activeSessionId]);
-
-  // Clear all storage
-  const clearAllStorage = useCallback(async (): Promise<void> => {
-    if (!storage) return;
-    try {
-      await storage.removeItem(keys.sessions);
-      await storage.removeItem(keys.activeSessionId);
-    } catch (err) {
-      console.error('Clear secure storage error:', err);
-    }
-  }, [storage, keys]);
-
-  // Switch to a different session
-  const switchToSession = useCallback(async (sessionId: string): Promise<void> => {
-    try {
-      setIsLoading(true);
-
-      // Get access token for this session
-      const tokenResponse = await oxyServices.getTokenBySession(sessionId);
-      // Set the token on the service instance
-      oxyServices.setTokens(tokenResponse.accessToken, '');
-
-      // Load full user data
-      const fullUser = await oxyServices.getUserBySession(sessionId);
-
-      setActiveSessionId(sessionId);
-      setUser(fullUser);
-      setMinimalUser({
-        id: fullUser.id,
-        username: fullUser.username,
-        avatar: fullUser.avatar
-      });
-
-      await saveActiveSessionId(sessionId);
-
-      if (onAuthStateChange) {
-        onAuthStateChange(fullUser);
-      }
-    } catch (error) {
-      console.error('Switch session error:', error);
-      setError('Failed to switch session');
-    } finally {
-      setIsLoading(false);
-    }
-  }, [oxyServices, onAuthStateChange, saveActiveSessionId]);
-
-  // Helper function to ensure token is set before API calls
-  const ensureToken = useCallback(async (): Promise<void> => {
-    console.log('ensureToken: Starting token check...');
-    console.log('ensureToken: Active session ID:', activeSessionId);
-
-    if (!activeSessionId) {
-      console.error('ensureToken: No active session ID found');
-      throw new Error('No active session');
-    }
-
-    try {
-      console.log('ensureToken: Getting token for session:', activeSessionId);
-      // Get the current token for the active session
-      const tokenResponse = await oxyServices.getTokenBySession(activeSessionId);
-      console.log('ensureToken: Token retrieved successfully:', !!tokenResponse.accessToken);
-
-      // Set the token on the service instance so it can be used for API calls
-      oxyServices.setTokens(tokenResponse.accessToken, ''); // No refresh token needed for session-based auth
-      console.log('ensureToken: Token set on service instance');
-
-    } catch (error) {
-      console.error('ensureToken: Failed to get token for session:', error);
-      throw new Error('Authentication failed. Please log in again.');
-    }
-  }, [activeSessionId, oxyServices]);
-
-  // Helper function to refresh user data from server
-  const refreshUserData = useCallback(async (): Promise<void> => {
-    if (!activeSessionId) {
-      console.error('refreshUserData: No active session ID found');
-      return;
-    }
-
-    try {
-      console.log('refreshUserData: Refreshing user data...');
-
-      // Ensure token is set
-      await ensureToken();
-
-      // Get fresh user data from server
-      const freshUser = await oxyServices.getCurrentUser();
-
-      // Update context state
-      setUser(freshUser);
-      setMinimalUser({
-        id: freshUser.id,
-        username: freshUser.username,
-        avatar: freshUser.avatar
-      });
-
-      console.log('refreshUserData: User data refreshed successfully');
-
-      if (onAuthStateChange) {
-        onAuthStateChange(freshUser);
-      }
-    } catch (error) {
-      console.error('refreshUserData: Failed to refresh user data:', error);
-      throw error;
-    }
-  }, [activeSessionId, oxyServices, ensureToken, onAuthStateChange]);
-
-  // Secure login method
-  const login = async (username: string, password: string, deviceName?: string): Promise<User> => {
-    if (!storage) throw new Error('Storage not initialized');
-
-    setIsLoading(true);
-    setError(null);
-
-    try {
-      // Get device fingerprint for enhanced device identification
-      const deviceFingerprint = DeviceManager.getDeviceFingerprint();
-
-      // Get or generate persistent device info
-      const deviceInfo = await DeviceManager.getDeviceInfo();
-
-      console.log('SecureAuth - Using device fingerprint:', deviceFingerprint);
-      console.log('SecureAuth - Using device ID:', deviceInfo.deviceId);
-
-      const response: SecureLoginResponse = await oxyServices.secureLogin(
-        username,
-        password,
-        deviceName || deviceInfo.deviceName || DeviceManager.getDefaultDeviceName(),
-        deviceFingerprint
-      );
-
-      // Create client session object with user info for duplicate detection
-      const clientSession: SecureClientSession = {
-        sessionId: response.sessionId,
-        deviceId: response.deviceId,
-        expiresAt: response.expiresAt,
-        lastActivity: new Date().toISOString(),
-        userId: response.user.id,
-        username: response.user.username
-      };
-
-      // Check if this user already has a session (prevent duplicate accounts)
-      const existingUserSessionIndex = sessions.findIndex(s =>
-        s.userId === response.user.id || s.username === response.user.username
-      );
-
-      let updatedSessions: SecureClientSession[];
-
-      if (existingUserSessionIndex !== -1) {
-        // User already has a session - replace it with the new one (reused session scenario)
-        const existingSession = sessions[existingUserSessionIndex];
-        updatedSessions = [...sessions];
-        updatedSessions[existingUserSessionIndex] = clientSession;
-
-        console.log(`Reusing/updating existing session for user ${response.user.username}. Previous session: ${existingSession.sessionId}, New session: ${response.sessionId}`);
-
-        // If the replaced session was the active one, update active session
-        if (activeSessionId === existingSession.sessionId) {
-          setActiveSessionId(response.sessionId);
-          await saveActiveSessionId(response.sessionId);
-        }
-      } else {
-        // Add new session for new user
-        updatedSessions = [...sessions, clientSession];
-        console.log(`Added new session for user ${response.user.username} on device ${response.deviceId}`);
-      }
-
-      setSessions(updatedSessions);
-      await saveSessionsToStorage(updatedSessions);
-
-      // Set as active session
-      setActiveSessionId(response.sessionId);
-      await saveActiveSessionId(response.sessionId);
-
-      // Get access token for API calls
-      const tokenResponse = await oxyServices.getTokenBySession(response.sessionId);
-      // Set the token on the service instance
-      oxyServices.setTokens(tokenResponse.accessToken, '');
-
-      // Load full user data
-      const fullUser = await oxyServices.getUserBySession(response.sessionId);
-      setUser(fullUser);
-      setMinimalUser(response.user);
-
-      if (onAuthStateChange) {
-        onAuthStateChange(fullUser);
-      }
-
-      return fullUser;
-    } catch (error: any) {
-      setError(error.message || 'Login failed');
-      throw error;
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  // Logout method
-  const logout = async (targetSessionId?: string): Promise<void> => {
-    if (!activeSessionId) return;
-
-    try {
-      const sessionToLogout = targetSessionId || activeSessionId;
-      await oxyServices.logoutSecureSession(activeSessionId, sessionToLogout);
-
-      // Remove session from local storage
-      const filteredSessions = sessions.filter(s => s.sessionId !== sessionToLogout);
-      setSessions(filteredSessions);
-      await saveSessionsToStorage(filteredSessions);
-
-      // If logging out active session
-      if (sessionToLogout === activeSessionId) {
-        if (filteredSessions.length > 0) {
-          // Switch to another session
-          await switchToSession(filteredSessions[0].sessionId);
-        } else {
-          // No sessions left
-          setActiveSessionId(null);
-          setUser(null);
-          setMinimalUser(null);
-          await storage?.removeItem(keys.activeSessionId);
-
-          if (onAuthStateChange) {
-            onAuthStateChange(null);
-          }
-        }
-      }
-    } catch (error) {
-      console.error('Logout error:', error);
-      setError('Logout failed');
-    }
-  };
-
-  // Logout all sessions
-  const logoutAll = async (): Promise<void> => {
-    console.log('logoutAll called with activeSessionId:', activeSessionId);
-
-    if (!activeSessionId) {
-      console.error('No active session ID found, cannot logout all');
-      setError('No active session found');
-      throw new Error('No active session found');
-    }
-
-    if (!oxyServices) {
-      console.error('OxyServices not initialized');
-      setError('Service not available');
-      throw new Error('Service not available');
-    }
-
-    try {
-      console.log('Calling oxyServices.logoutAllSecureSessions with sessionId:', activeSessionId);
-      await oxyServices.logoutAllSecureSessions(activeSessionId);
-      console.log('logoutAllSecureSessions completed successfully');
-
-      // Clear all local data
-      setSessions([]);
-      setActiveSessionId(null);
-      setUser(null);
-      setMinimalUser(null);
-      await clearAllStorage();
-      console.log('Local storage cleared');
-
-      if (onAuthStateChange) {
-        onAuthStateChange(null);
-        console.log('Auth state change callback called');
-      }
-    } catch (error) {
-      console.error('Logout all error:', error);
-      setError(`Logout all failed: ${error instanceof Error ? error.message : 'Unknown error'}`);
-      throw error;
-    }
-  };
-
-  // Sign up method
-  const signUp = async (username: string, email: string, password: string): Promise<User> => {
-    if (!storage) throw new Error('Storage not initialized');
-
-    setIsLoading(true);
-    setError(null);
-
-    try {
-      // Create new account using the OxyServices signUp method
-      const response = await oxyServices.signUp(username, email, password);
-
-      console.log('SignUp successful:', response);
-
-      // Now log the user in securely to create a session
-      // This will handle the session creation and device registration
-      const user = await login(username, password);
-
-      return user;
-    } catch (error: any) {
-      setError(error.message || 'Sign up failed');
-      throw error;
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  // Switch session method
-  const switchSession = async (sessionId: string): Promise<void> => {
-    await switchToSession(sessionId);
-  };
-
-  // Remove session method
-  const removeSession = async (sessionId: string): Promise<void> => {
-    await logout(sessionId);
-  };
-
-  // Refresh sessions method
-  const refreshSessions = async (): Promise<void> => {
-    if (!activeSessionId) return;
-
-    try {
-      const serverSessions = await oxyServices.getSessionsBySessionId(activeSessionId);
-
-      // Update local sessions with server data
-      const updatedSessions: SecureClientSession[] = serverSessions.map(serverSession => ({
-        sessionId: serverSession.sessionId,
-        deviceId: serverSession.deviceId,
-        expiresAt: new Date().toISOString(), // You might want to get this from server
-        lastActivity: new Date().toISOString()
-      }));
-
-      setSessions(updatedSessions);
-      await saveSessionsToStorage(updatedSessions);
-    } catch (error) {
-      console.error('Refresh sessions error:', error);
-    }
-  };
-
-  // Device management methods
-  const getDeviceSessions = async (): Promise<any[]> => {
-    if (!activeSessionId) throw new Error('No active session');
-
-    try {
-      return await oxyServices.getDeviceSessions(activeSessionId);
-    } catch (error) {
-      console.error('Get device sessions error:', error);
-      throw error;
-    }
-  };
-
-  const logoutAllDeviceSessions = async (): Promise<void> => {
-    if (!activeSessionId) throw new Error('No active session');
-
-    try {
-      await oxyServices.logoutAllDeviceSessions(activeSessionId);
-
-      // Clear all local sessions since we logged out from all devices
-      setSessions([]);
-      setActiveSessionId(null);
-      setUser(null);
-      setMinimalUser(null);
-      await clearAllStorage();
-
-      if (onAuthStateChange) {
-        onAuthStateChange(null);
-      }
-    } catch (error) {
-      console.error('Logout all device sessions error:', error);
-      throw error;
-    }
-  };
-
-  const updateDeviceName = async (deviceName: string): Promise<void> => {
-    if (!activeSessionId) throw new Error('No active session');
-
-    try {
-      await oxyServices.updateDeviceName(activeSessionId, deviceName);
-
-      // Update local device info
-      await DeviceManager.updateDeviceName(deviceName);
-    } catch (error) {
-      console.error('Update device name error:', error);
-      throw error;
-    }
-  };
-
-  // Bottom sheet control methods
-  const showBottomSheet = useCallback((screenOrConfig?: string | { screen: string; props?: Record<string, any> }) => {
-    console.log('showBottomSheet called with:', screenOrConfig);
-
-    if (bottomSheetRef?.current) {
-      console.log('bottomSheetRef is available');
-
-      // First, show the bottom sheet
-      if (bottomSheetRef.current.expand) {
-        console.log('Expanding bottom sheet');
-        bottomSheetRef.current.expand();
-      } else if (bottomSheetRef.current.present) {
-        console.log('Presenting bottom sheet');
-        bottomSheetRef.current.present();
-      } else {
-        console.warn('No expand or present method available on bottomSheetRef');
-      }
-
-      // Then navigate to the specified screen if provided
-      if (screenOrConfig) {
-        // Add a small delay to ensure the bottom sheet is opened first
-        setTimeout(() => {
-          if (typeof screenOrConfig === 'string') {
-            // Simple screen name
-            console.log('Navigating to screen:', screenOrConfig);
-            bottomSheetRef.current?._navigateToScreen?.(screenOrConfig);
-          } else {
-            // Screen with props
-            console.log('Navigating to screen with props:', screenOrConfig.screen, screenOrConfig.props);
-            bottomSheetRef.current?._navigateToScreen?.(screenOrConfig.screen, screenOrConfig.props);
-          }
-        }, 100);
-      }
-    } else {
-      console.warn('bottomSheetRef is not available');
-    }
-  }, [bottomSheetRef]);
-
-  const hideBottomSheet = useCallback(() => {
-    if (bottomSheetRef?.current) {
-      bottomSheetRef.current.dismiss?.();
-    }
-  }, [bottomSheetRef]);
-
-  // Compute comprehensive authentication status
-  // This is the single source of truth for authentication across the entire app
-  const isAuthenticated = useMemo(() => {
-    // User is authenticated if:
-    // 1. We have a full user object loaded, OR
-    // 2. We have an active session with a valid token
-    // This covers both the loaded state and the loading-but-authenticated state
-    return !!user || (!!activeSessionId && !!oxyServices?.getCurrentUserId());
-  }, [user, activeSessionId, oxyServices]);
-
-  // Context value
-  const contextValue: OxyContextState = {
-    user,
-    minimalUser,
-    sessions,
-    activeSessionId,
-    isAuthenticated,
-    isLoading,
-    error,
-    login,
-    logout,
-    logoutAll,
-    signUp,
-    switchSession,
-    removeSession,
-    refreshSessions,
-    getDeviceSessions,
-    logoutAllDeviceSessions,
-    updateDeviceName,
-    oxyServices,
-    bottomSheetRef,
-    showBottomSheet,
-    hideBottomSheet,
-    ensureToken,
-    refreshUserData,
-  };
+  }, [oxyServices]);
 
   return (
-    <OxyContext.Provider value={contextValue}>
+    <OxyContext.Provider value={{ 
+      oxyServices, 
+      bottomSheetRef, 
+      showBottomSheet, 
+      hideBottomSheet 
+    }}>
+      <AuthStateChangeListener onAuthStateChange={onAuthStateChangeRef.current} />
       {children}
     </OxyContext.Provider>
   );
 };
 
-// Hook to use the context
-export const useOxy = (): OxyContextState => {
-  const context = useContext(OxyContext);
-  if (!context) {
-    throw new Error('useOxy must be used within an OxyContextProvider');
-  }
-  return context;
+/**
+ * Component to handle auth state change callbacks for backward compatibility
+ */
+const AuthStateChangeListener: React.FC<{ onAuthStateChange?: (user: User | null) => void }> = ({ 
+  onAuthStateChange 
+}) => {
+  const auth = useAuth();
+  
+  useEffect(() => {
+    if (onAuthStateChange) {
+      onAuthStateChange(auth.user);
+    }
+  }, [auth.user, onAuthStateChange]);
+
+  return null;
 };
 
-export default OxyContext;
+/**
+ * Hook to access the OxyContext
+ */
+export const useOxyContext = (): OxyContextState => {
+  const context = useContext(OxyContext);
+  if (!context) {
+    throw new Error('useOxyContext must be used within an OxyContextProvider');
+  }
+  
+  const auth = useAuth();
+  const follow = useFollow();
+
+  return {
+    // Context values
+    ...context,
+    
+    // Authentication state and actions
+    user: auth.user,
+    minimalUser: auth.minimalUser,
+    sessions: auth.sessions,
+    activeSessionId: auth.activeSessionId,
+    isAuthenticated: auth.isAuthenticated,
+    isLoading: auth.isLoading,
+    error: auth.error,
+
+    // Auth actions
+    login: auth.login,
+    logout: auth.logout,
+    logoutAll: auth.logoutAll,
+    signUp: auth.signUp,
+    switchSession: auth.switchSession,
+    removeSession: auth.removeSession,
+    refreshSessions: auth.refreshSessions,
+    refreshUserData: auth.refreshUserData,
+    
+    // Device management
+    getDeviceSessions: auth.getDeviceSessions,
+    logoutAllDeviceSessions: auth.logoutAllDeviceSessions,
+    updateDeviceName: auth.updateDeviceName,
+    ensureToken: auth.ensureToken,
+  };
+};
+
+/**
+ * Main hook that combines all Oxy functionality
+ * This is the primary hook that components should use
+ */
+export const useOxy = () => {
+  const context = useOxyContext();
+  const follow = useFollow();
+
+  return {
+    // All context functionality
+    ...context,
+
+    // Follow functionality
+    followingUsers: follow.followingUsers,
+    loadingUsers: follow.loadingUsers,
+    followErrors: follow.errors,
+    
+    // Follow actions
+    toggleFollow: follow.toggleFollow,
+    followUser: follow.followUser,
+    unfollowUser: follow.unfollowUser,
+    fetchFollowStatus: follow.fetchFollowStatus,
+    fetchMultipleStatuses: follow.fetchMultipleStatuses,
+    setFollowingStatus: follow.setFollowingStatus,
+    clearFollowError: follow.clearFollowError,
+    clearAllFollowErrors: follow.clearAllFollowErrors,
+  };
+};
+
+// Export for legacy support - keeping the original interface
+export const OxyProvider = OxyContextProvider;
+
+// Export types
+export type { OxyContextState, OxyContextProviderProps };
