+++ packages/services/src/ui/hooks/index.ts
@@ -1,1 +1,2 @@
-export { useOxyFollow, useFollow } from './useOxyFollow'; 
+export { useFollow, useFollowUser, useFollowMultipleUsers } from './useFollow';
+export { useOxyFollow } from './useOxyFollow';
